Homework2/src/bot/requirements.txt
requests
piexif
pillow